setup for tests
