+++ package.json
@@ -6,7 +6,6 @@
   "contributors": [
     "Erhan Tezcan <erhan@firstbatch.xyz> (https://github.com/erhant)"
   ],
-  "homepage": "https://github.com/firstbatchxyz/dria-cli#readme",
   "license": "Apache-2.0",
   "files": [
     "/bin",
@@ -17,12 +16,12 @@
     "dria": "./bin/index.js"
   },
   "scripts": {
-    "lint": "npx eslint './src/**/*.ts' && echo 'All good!'",
-    "format": "npx prettier --check ./src/**/*.ts",
-    "build": "npx tsc",
+    "build": "npx tsc -p tsconfig.build.json",
     "start": "node ./bin/index.js",
     "dria": "yarn build && yarn start",
-    "test": "npx jest"
+    "test": "npx jest",
+    "lint": "npx eslint './src/**/*.ts' && echo 'All good!'",
+    "format": "npx prettier --check ./src/**/*.ts"
   },
   "engines": {
     "node": ">=18.0.0"
@@ -49,6 +48,7 @@
     "jest": "^29.7.0",
     "prettier": "^3.2.5",
     "ts-jest": "^29.1.2",
+    "ts-node": "^10.9.2",
     "typescript": "^5.3.3"
   },
   "prettier": {
@@ -65,6 +65,27 @@
     ],
     "rules": {
       "@typescript-eslint/no-unused-vars": "warn"
-    }
-  }
+    },
+    "ignorePatterns": [
+      "bin",
+      "node_modules"
+    ]
+  },
+  "homepage": "https://github.com/firstbatchxyz/dria-cli#readme",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/firstbatchxyz/dria-cli.git"
+  },
+  "bugs": {
+    "url": "https://github.com/firstbatchxyz/dria-cli/issues/"
+  },
+  "keywords": [
+    "dria",
+    "hollowdb",
+    "arweave",
+    "yargs",
+    "cli",
+    "rag",
+    "ai"
+  ]
 }
