+++ tests/commands.test.ts
@@ -1,0 +1,5 @@
+describe.skip("commands", () => {
+  it("should do something", () => {
+    console.log("hi");
+  });
+});
