--- tests/index.test.ts
@@ -1,3 +1,0 @@
-describe("cli tests", () => {
-  // TODO: https://kgajera.com/blog/how-to-test-yargs-cli-with-jest/
-});
