+++ tsconfig.build.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "bin",
+    "noEmit": false
+  },
+  "include": ["src/**/*.ts"]
+}
